# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Routeyou events API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/organisations": {
            "get": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Get all organisations",
                "operationId": "adf4f1a323bdec2d8709835a3d3ef653",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved organisations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The organisations are returned successfully."
                                        },
                                        "organisations": {
                                            "example": "Array with the organisations will be returned here."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Create a new organisation and link to user",
                "operationId": "d2dd14efb9f5b86525039c766e1f2442",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Organisation"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A brief description of the organisation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Organisation created successfully and linked to the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The organisation has been created successfully and linked to your account."
                                        },
                                        "organisation": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "New Organisation",
                                                "description": "A brief description of the organisation"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not allowed to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "description": [
                                                    "The description must be a string."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": [
                            "USER"
                        ]
                    }
                ]
            }
        },
        "/api/organisations/{organisation}": {
            "get": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Get a specific organisation",
                "operationId": "f17dcee39bec153e196814d7a4067cb8",
                "parameters": [
                    {
                        "name": "organisation",
                        "in": "path",
                        "description": "ID of the organisation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the organisation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The organisation has been returned successfully."
                                        },
                                        "organisation": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Test Organisation",
                                                "description": "This is an example description",
                                                "organisers": "Array",
                                                "followers": "Array"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organisation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Organisation]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Delete an organisation",
                "operationId": "a63f6f9b3081e80a0902e8df181fad5c",
                "parameters": [
                    {
                        "name": "organisation",
                        "in": "path",
                        "description": "ID of the organisation to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the organisation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The organisation has been deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not allowed to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organisation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Organisation]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": [
                            "ORGANISER"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}